Organized python FUT into separate folder
