+++ examples/exploring_bnn_perf/client/data.py
