+++ examples/exploring_bnn_perf/client/network.py
