+++ examples/exploring_bnn_perf/client/sembas_api.py
